Report correct time instead of old time when locating as Jack transport master.
I encountered this today writing a transport slave, but it seems to be the same
problem as issue #743 from 8 years ago.  The issue is easier to see with a
transport slave client that prints any transport change whatsoever, for example
if the current location is some point well into the session and rewind to start
is pressed, Ardour reports the old position, not zero.  With this change, it
reports zero as expected.

See comment about why this was happening.  If locating here is evil for some
reason, then some other way of making jack_timebase_callback report the target
position is required.  Contrary to what the old comment below this change
suggests, follow_slave() does not update _transport_frame in time.


git-svn-id: svn://localhost/ardour2/branches/3.0@12993 d708f5d6-7413-0410-9779-e7cbd77b26cf