disable symbol export -- fix LinuxVST issue
-rdynamic:  Pass the flag -export-dynamic to the ELF linker,
 on targets that support it. This instructs the linker to add
 all symbols, not only used ones, to the dynamic symbol table.
 This option is needed for some uses of dlopen or to allow
obtaining backtraces from within a program.

Some LinuxVST pluging cause ardour3 to crash -- if ardour is linked with
-rdynamic (see #ardour log from 2013-jun-6). This is likely compiler
specific and not a problem on las' build-machine (gcc 4.5.5) which is
why the linux-packaging script enables it by default.
The issue can be reproduced reliably with gcc 4.7.2 and 4.6.3. Although
it is yet unclear if it is caused by g++/ld or related to other factors
of the build-system...