vtl: check if server's docroot matches ardour's config
NOTE: this breaks backwards compatibility with icsd and
harvid < 0.3.0 which do not report their settings.

It may also conflict with harvid running on localhost in a chroot.

Ideally this will be a warning only.
That the user can choose to ignore this and should get the option
to stop the video-server and re-start it using a different docroot.