keep processing LV2 Ringbuffer once it is allocated
Discard messages in LV2PluginUI::write_to_ui() if the
ringbuffer exists but the UI is not active.

This fixes "[ERROR]: Error writing from plugin to UI" messages
if the UI was once active but has been closed since.

This is a hotfix solution, eventually the LV2 backend should be updated
and a function LV2Plugin::disable_ui_emmission() implemented.