Fix windows key bindings.
I am not sure if this is correct, but since the name in the bindings didn't
match that in keyboard.cc, the modifier was just ignored, which results in
conflicting bindings and consequent bugs.

"Meta" seems to be the name for this thing (oddly, to me), so I chose that,
though since my window manager steals this key (as basically all the standard
ones do these days) I can't tell if these bindings actually work with the
desired key.