Revert "don't unpack and repack meters unnecessarily during LevelMeterBase::setup_meters()"
This reverts commit d00e035fe0618c8e3f4c87c6386fc87efb58cd62.

..which introduced a bug.

Meter count and type may change when switching the meter point or simply
when selecting a different track in the editor-mixer. Additional meters
were not unpacked and remained on canvas.

Ideally we'd just remove the excess meters or meters where the type
changed, but that's far more complex than just repacking them all.